% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecheckbox.R
\name{treecheckbox}
\alias{treecheckbox}
\title{Treecheckbox}
\usage{
treecheckbox(
  id,
  label,
  choices,
  levels = c(),
  collapsed = FALSE,
  selected = NULL,
  include_mode = FALSE,
  select_buttons = TRUE,
  search_bar = TRUE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{id}{The \code{input} slot that will be used to access the value.}

\item{label}{The label that will be displayed in front. If \code{NULL} is provided, no label will be generated.}

\item{choices}{A dataframe which will be used to generate the checkboxes.}

\item{levels}{A vector which contains the columnames that will be used to create the hierarchical data. From large to small groups.}

\item{collapsed}{Logical; If \code{TRUE} checkboxes will be collapsed on render.}

\item{selected}{A vector containing the values of which checkboxes will default as checked.}

\item{include_mode}{Logical; If \code{TRUE} the checkboxes will be rendered with include/exclude mode.}

\item{select_buttons}{Logical; If \code{TRUE} helper buttons will be rendered to select all or none, collapse or expand.}

\item{search_bar}{Logical; If \code{TRUE} a search bar will be rendered}
}
\description{
Creates the checkboxes.
}
\examples{
library(shiny)
library(shinytreecheckbox)
choices = data.frame(order = c("Primates", "Primates", "Primates", "Primates", "Primates", "Carnivora", "Carnivora", "Birds", "Birds", "Birds", "Fish", "Fish", "Fish"), genus = c("Baboons", "Capuchin monkeys", "Chimpanzees", "Gorillas", "Mandrills", "Seals", "Candids", "Chiroxiphia", "Montezuma oropendolas", "Pale chanting goshawks", "Cichlids", "Moon wrasse", "Mozambique tilapia"))
ui <- fluidPage(
 shinytreecheckbox::treecheckbox("thisisanid", "mytreecheckbox", choices, c("order", "genus"), T, c("Fish", "Gorillas"))
)
server <- function(input, output, session) {}
shinyApp(ui, server)

}
