% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecheckbox.R
\name{treecheckbox}
\alias{treecheckbox}
\title{Treecheckbox}
\usage{
treecheckbox(
  id,
  label,
  choices,
  levels,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
Creates the checkboxes.
}
